close: check for nil cmd before trying to kill it
Check for a nil cmd before trying to access it's memory. If a cmd exits
because the underlying process closed (as opposed to being closed via a
call to Close), the cmd reference can be nil when we try to kill it.

+++ vt.go
@@ -290,8 +290,10 @@ func (vt *VT) Start(cmd *exec.Cmd) error {
 func (vt *VT) Close() {
 	vt.mu.Lock()
 	defer vt.mu.Unlock()
-	vt.cmd.Process.Kill()
-	vt.cmd.Wait()
+	if vt.cmd != nil {
+		vt.cmd.Process.Kill()
+		vt.cmd.Wait()
+	}
 	vt.pty.Close()
 }
